Sep 2, 2024, 9:32 PM

+++ cos.c
@@ -8,7 +8,7 @@
 
     int main(void)
     {
-        long x, y, x_sq, s = 0.0, t, prev;
+        float x, y, x_sq, s = 0.0, t, prev;
         int i = 0;
 
         // Input value in radians
